src/main/java/com/openclassrooms/safetynet/utils/DateUtils.java
package com.openclassrooms.safetynet.utils;

import java.time.LocalDate;
import java.time.Period;
import java.time.format.DateTimeFormatter;

public class DateUtils {
    public static int calculateAge(String birthdate) {
        if (birthdate == null) {
            return 0;
        }

        LocalDate birthDate = LocalDate.parse(birthdate, DateTimeFormatter.ofPattern("MM/dd/yyyy"));
        LocalDate currentDate = LocalDate.now();
        return Period.between(birthDate, currentDate).getYears();
    }
}
